+++ HSTracker/Logging/Enums/GameTag.swift
@@ -412,6 +412,7 @@ enum GameTag: Int, CaseIterable, Codable {
     gametag_2358 = 2358,
     was_discover_option = 2509,
     bacon_buddy_enabled = 2518,
+    bacon_evolution_card_id = 2519,
     immolatestage = 2600,
     corpses_spent_this_game = 2639,
     bacon_card_dbid_reward = 2673,
@@ -509,6 +510,7 @@ enum GameTag: Int, CaseIterable, Codable {
     hidden_script_data_4 = 4830,
     hidden_script_data_5 = 4831,
     hidden_script_data_6 = 4832,
+    dark_gift_entity = 4865,
     bacon_activate_tooltip = 4867
 
     static var lookup = [String: GameTag]()